deprecate the interactive mode and add a better cli
