+++ instabot_py/__init__.py
@@ -1,3 +1,5 @@
 from .instabot import InstaBot
 
+VERSION = '0.5.0'
 __all__ = ['InstaBot']
+__version__ = VERSION
