+++ requirements/ci.txt
@@ -1,2 +1,1 @@
 tox
-pytest-cov
